rename the client service
