+++ whois-api/src/main/java/net/ripe/db/whois/api/rest/SyncUpdatesService.java
@@ -11,7 +11,7 @@ import net.ripe.db.whois.common.conversion.PasswordFilter;
 import net.ripe.db.whois.common.domain.CIString;
 import net.ripe.db.whois.common.source.SourceContext;
 import net.ripe.db.whois.common.sso.AuthServiceClient;
-import net.ripe.db.whois.common.sso.CrowdClientException;
+import net.ripe.db.whois.common.sso.AuthServiceClientException;
 import net.ripe.db.whois.common.sso.SsoTokenTranslator;
 import net.ripe.db.whois.update.domain.ContentWithCredentials;
 import net.ripe.db.whois.update.domain.Keyword;
@@ -214,7 +214,7 @@ public class SyncUpdatesService {
         if (!StringUtils.isBlank(ssoToken)) {
             try {
                 updateContext.setUserSession(ssoTokenTranslator.translateSsoToken(ssoToken));
-            } catch (CrowdClientException e) {
+            } catch (AuthServiceClientException e) {
                 loggerContext.log(new Message(Messages.Type.ERROR, e.getMessage()));
                 updateContext.addGlobalMessage(RestMessages.ssoAuthIgnored());
             }
