--- whois-client/src/main/java/net/ripe/db/whois/common/sso/CrowdClientException.java
@@ -1,27 +1,0 @@
-package net.ripe.db.whois.common.sso;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.ws.rs.ClientErrorException;
-import javax.ws.rs.ProcessingException;
-import javax.ws.rs.WebApplicationException;
-
-public class CrowdClientException extends RuntimeException {
-
-    private final String message;
-    private final int code;
-
-    public CrowdClientException(int code, String message) {
-        this.code = code;
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public int getCode() {
-        return code;
-    }
-}
